Add PyMySQL to optional dependencies
A library called [PyMySQL] provides utilities for converting Python
values into MySQL literals, which comes in very handy now that we want
to add MySQL support into the sanitation tool. It differs from other
MySQL libraries for Python so that it's pure Python and does not require
compilation of C libraries, so I chose it for the task.

Also add PyMySQL as non-optional dependency for test requirements, or
otherwise test won't be properly run.

[PyMySQL]: https://github.com/PyMySQL/PyMySQL