Add sanitation sesssion API
Sanitation session allows having a state within a single sanitation
process.

One important thing stored to the session is a secret key which is
generated to a new random value for each sanitation session, but it
stays constant during the whole sanitation process. Its value is never
revealed, so that it is possible to generate such one way hashes with
it, that should not be redoable afterwards. I.e. during the sanitation
session it's possible to do ``hash(C) -> H`` for any clear text C, but
it is not possible to check if H is the hashed value of C after the
sanitation session has ended.